requirements.txt
web3
eth-brownie
